requirements.txt
aioboto3==10.0.0; python_version >= "3.7" and python_version < "4.0"
aiobotocore==2.3.4; python_version >= "3.7" and python_version < "4.0"
aiohttp==3.8.1; python_version >= "3.6"
aioitertools==0.10.0; python_version >= "3.7" and python_version < "4.0"
aiosignal==1.2.0; python_version >= "3.7" and python_version < "4.0"
anyio==3.6.1
asgiref==3.5.2; python_full_version >= "3.6.1" and python_version >= "3.7"
async-timeout==4.0.2; python_version >= "3.7" and python_version < "4.0"
attrs==22.1.0; python_version >= "3.7" and python_version < "4.0"
beautifulsoup4==4.11.1; python_full_version >= "3.6.0"
boto3==1.21.21; python_version >= "3.7" and python_version < "4.0"
botocore==1.24.21; python_version >= "3.7" and python_version < "4.0"
bs4==0.0.1
certifi==2022.6.15; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.7.0"
charset-normalizer==2.1.1; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "4"
click==8.1.3; python_version >= "3.7" and python_full_version >= "3.7.0"
colorama==0.4.5
dnspython==2.2.1; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.1"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
email-validator==1.2.1; python_full_version >= "3.6.1"
fastapi==0.81.0; python_full_version >= "3.6.1"
faunadb==4.2.0
frozenlist==1.3.1; python_version >= "3.7" and python_version < "4.0"
future==0.18.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
h11==0.12.0; python_full_version >= "3.7.0" and python_version >= "3.7"
h2==2.6.2
hpack==3.0.0
httpcore==0.15.0; python_version >= "3.7" and python_full_version >= "3.7.0"
httptools==0.4.0
httpx==0.23.0; python_version >= "3.7" and python_full_version >= "3.7.0"
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iso8601==1.0.2; python_full_version >= "3.6.2" and python_version < "4.0"
itsdangerous==2.1.2; python_full_version >= "3.6.1" and python_version >= "3.7"
jinja2==3.1.2; python_full_version >= "3.7.0" and python_version >= "3.7"
jmespath==1.0.1; python_version >= "3.7" and python_version < "4.0"
lxml==4.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
markupsafe==2.1.1; python_full_version >= "3.7.0" and python_version >= "3.7"
multidict==6.0.2; python_version >= "3.7" and python_version < "4.0"
orjson==3.8.0; python_full_version >= "3.6.1" and python_version >= "3.7"
prisma==0.6.6; python_full_version >= "3.7.0"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pydantic==1.9.2; python_full_version >= "3.7.0"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5; python_full_version >= "3.6.1"
pytube==12.1.0; python_version >= "3.6"
pyyaml==6.0
requests==2.28.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1"
rfc3986==1.5.0; python_version >= "3.7" and python_full_version >= "3.7.0"
rsa==4.9; python_version >= "3.6" and python_version < "4"
s3transfer==0.5.2; python_version >= "3.7" and python_version < "4.0"
six==1.16.0; python_full_version >= "3.6.1" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0") and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0")
sniffio==1.2.0; python_version >= "3.7" and python_full_version >= "3.7.0"
soupsieve==2.3.2.post1; python_version >= "3.6" and python_full_version >= "3.6.0"
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.3.0; python_version >= "3.7" and python_full_version >= "3.7.0" and python_version < "3.10"
ujson==5.4.0; python_full_version >= "3.6.1" and python_version >= "3.7"
urllib3==1.26.12; python_full_version >= "3.6.1" and python_version < "4" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7")
uvicorn==0.17.6; python_full_version >= "3.6.1" and python_version >= "3.7"
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
yarl==1.8.1; python_version >= "3.7" and python_version < "4.0"
